Move `DelayIter` to dedicated module
