+++ src/lib.rs
@@ -27,6 +27,7 @@
 #![deny(missing_docs, broken_intra_doc_links)]
 
 pub mod flat;
+pub mod iter;
 pub mod trapezoidal;
 pub mod util;
 
@@ -83,23 +84,7 @@ pub trait MotionProfile: Sized {
     ///
     /// This is a convenience method that returns an iterator which internally
     /// just calls [`MotionProfile::next_delay`].
-    fn delays(&mut self) -> DelayIter<Self> {
-        DelayIter(self)
-    }
-}
-
-/// An iterator over delay values
-///
-/// Can be created by calling [`MotionProfile::delays`].
-pub struct DelayIter<'r, T>(pub &'r mut T);
-
-impl<'r, T> Iterator for DelayIter<'r, T>
-where
-    T: MotionProfile,
-{
-    type Item = T::Delay;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next_delay()
+    fn delays(&mut self) -> iter::DelayIter<Self> {
+        iter::DelayIter(self)
     }
 }
